TacRP Support & shared CVars
Moved CVar definitions to shared and added replicated to some CVars. This fixes a lua error that spams the console when opening the options menu on dedicated servers (though this menu probably needs to be hidden or restricted to superadmins on those servers anyway. That's for later).

Added hardcoded support for tacrp_base as a bunch of people requested that.