Better TacRP integration, define GM Name/Author, More reliable(?) disable of ai_serverragdolls
Added a hook to remove the invasive HUD elements defined in TacRP (whilst keeping the customisation menu). Also found the bool marking weapons as from that base and replaced the check in cost calc with that.

Made it so that the console-command-run to disable ai_serverragdolls runs and additional 2 times, and also called game.ConsoleCommand in-case RunConsoleCommand isn't working for some reason. I don't know why the first one *sometimes* fails, or if this will fix that, but I don't know what else to do about it. People still have this issue sometimes.

Defined GM.Name and GM.Author. Maybe this is why it shows up with no name in the server browser?