Sort-of Fixes/better errors
Payload now exits an attempt if it can't get an area. This implies the map is unsuitable/it won't work anyway, but I'd like the error to actually *say* that so people aren't confused.

Added an error for cases where .Primary doesn't exist in a weapon's definition. Our code relies on/assumes this exists. If it isn't there then other parts of the code error. Also included the classname and base so that we can figure out what weapon actually broke it.