Error fixes & new player difficulty
IsValid checks instead of just checking if a panel exists. This errored with null panels.

Reduced the new-player difficulty further and increased the mission-completions required for it to hit normal level by 1.

Tried to reduce chances of pre-game/spectator players sort-of falling out of the map and spamming the console with engine errors. They're stacked on top of the first spawn point now.