Fixed npcs under displacements
Made it so that the tracehull for non-static enemies ends at the position the npc was placed, not where it started. This is probably almost redundant now but I'll worry about that later.

Made the straggler system place npcs 15u above the vector given by squadVectors, not on top of it. This appeared to be causing straggling npcs to be placed in positions where they would fall under the map. This also explains why I never saw this happening first-hand. I tested this by storing an npc's spawn pos and checking it whenever I found one under the map (which consistently showed them to have been spawned on the opposite end of the map, in positions where they wouldn't have fallen under).

Marked polyps and spawners as static enemies.

Stricter exclusion for squadVecs (this wasn't the issue, but it should ignore sky anyway.