Fix payload terminals falling beneath displacements
Replaced payload terminals' failsafe algorithm with one that uses navareas instead of arbitrary points to hopefully ensure they stay in the map, even in cramped spaces.

Removed :DropToFloor() call on the terminal. Replaced it with a custom function that snaps it to the navmesh instead.

Removed :DropToFloor calls in other areas of the code where possible. This function is incredibly unreliable and had already been removed in other parts of the code previously for this reason. This should hopefully prevent some other issues with things falling under the map in the future.