Minor fixes
Try to read node data whenever it's needed if it doesn't exist, as an extra failsafe. A few people have still reported errors relating to this even after the previous change. If this doesn't work that means the file literally doesn't exist sometimes, in-which case I need to do something totally different (I should probably try to figure out if the method Strawwagen mentioned works).

Fixed pathfinder potentially erroring if there are no nodes within the PVS of the start/end.