Fix issues with weapon price saving, orbital beam tweaks
jcms_deathray now has separate ENT.DPS and ENT.DPS_DIRECT values, affecting the direct damage (findalongray) and blast damage separately.

The spawn-menu orbital beam call-in has been modified to have substantially higher direct-damage whilst keeping the blast damage the same (in-order to make it a more effective boss-killer).
The beam-speed of the spawnmenu orbital beam has also been slightly increased.

Orbital beams have been changed to prioritse the highest max health entity, not the highest health entity. They would frequently bugger-off from whatever helicopter swarm or boss you placed them on, this should hopefully fix that (from testing it seems to, but hard to say without releasing it into the wild).

Stopped weapon prices being set to nil if the weapon is disabled. This is entirely unnecessary, as every other part of the code treats 0-cost as disabled, and doing this breaks saving of disabled weapons.
Removed jcms.weapon_prices_wereModified check from the weapon-saving hook, as this value does not reliably exist in this hook (I suspect various variables are not reliably present in this hook, which is a major concern.)