Include stack-trace in director pcall errors
Stack traces are now included in the non-halting-error thrown by director's pcall. Having only the error with no additional context made debugging more difficult, and sometimes impossible.
Recently CFC had an error that I could not identify, as it seems to have happened in an external function (the variable name referenced literally did not exist anywhere in the project). This should hopefully allow me to fix that/other issues in the future.