Lua error fix?
CFC was giving me errors earlier due to the TakePrimaryAmmo function not existing clientside. I wasn't able to reproduce this on singleplayer, p2p, or a locally hosted dedicated server (or while using their custom version of m9k). So I can't really confirm if this doesn't lead to other issues. In theory it should just result in slight desync and cause no errors, which is an improvement.

Also stopped the sounds played from being duplicated on both client and server now that this code is shared.