Fixes & More granular debris cleanup logic
Fixed potential null-player error in sentinel teleport

Fixed potential engine hard-crash due to the minimum distance to goal check reducing to 0 on high player counts for slow zombies. If an enemy has no path calling :AdvancePath() causes a crash, this is otherwise prevented by checking if path length is above 0.

Made item_battery and item_ammo_ar2_altfire follow the same logic as weapons in director debris cleanup, as they are important entities. 10s cleanup w/no delay for nearby sweepers means these are highly likely to be lost (and while in-vision of players) which would be bad.

Increased the time for heli-debris clearing up again, but only to 60s. I still want heli debris to be usable as cover, but I recognise why this change was made, and that having it remain indefinitely while players are present is incompatible with unblocking payloads, which take priority.