Brush data reading & better player-out-of-bounds engine-error prevention
Updated sh_bspreader.lua to allow reading of brush data (including brushes and brush faces). It still needs to be verified that this is correct.
I plan to use this in the future to read-in playerclips & detect whether or not they intersect navareas, in-order to mark them as invalid. This should hopefully allow maps with invisible walls to remain playable, and not place players or objectives in blocked-off areas.

Increased the jankiness of the code used to prevent spectating players from falling out of the world and causing engine errors. It now uses airgraph nodes if they exist, which should be more reliable. I may in the future want to pre-calculate a bunch of points based on the same visleaf data and use those instead, as using airgraph means this doesn't work on indoor maps (whereas a proper solution could work on any playable map).