Fix cfc/zmod m9k vers calculating RPG costs wrong
This version introduces custom ammo types (for some reason) which messes with cost calculations and ammo prices. This corrects that.

There's still something funky about this version that I need to figure out how to account for. Broadly weapons are cheaper than they should be because they have a different way of setting up spread. This is why most of the guns on those servers are half price or lower.