Scale infestation storm duration w/map characteristics
The duration of infestation's storms now scales down if the map is incredibly large and open. Blood-rain on maps like DIPRIP is generally boring, as it mostly consists of sitting under an umbrella and waiting for 90 seconds. I may need to tweak this a little in the future (maybe ^2/3 instead of sqrt) but it should help for now. Dense maps remain the same, as they are what it was designed around. Storm duration will only decrease under this change, not increase.

Approximate new duration as % of the total/default:
construct: 100%
jcorpdistrict: 100%
diprip_refinery: 70% --(might need to be lower)
gm_flatgrass: 30%