Mapgen progress / standardised yielding
the if coroutine.running() then coroutine.yield() end code used to pause during mapgen is now standardised. The function for this also contains a single argument for generation progress, which I have defined (approximately) for each mission generation type. This is eventually intended to be used in a loading screen.