Lobby NPC-Team warning.
The "Join as NPC" button in the lobby now displays a disclaimer / confirm pop-up window.

This ensures (if they bother to read it) that they first know it is unfair, and that they can't change back from it. I have seen several people choose it without understanding what they were signing up for, and then becoming frustrated with their lack of impact. This should hopefully prevent that in the future.