Infestation Custom Evac placement logic
Evacs on infestation missions will now try to spawn outside of the range of any radiation spheres that INCOMPETENT EMPLOYEES have left behind. This might at least partly be a dedicated server thing, but this has tended to be a low point for a lot of people, which I'd like to reduce the frequency of.