HUD Optimisation
Cached ScrW / ScrH and used the cached value in HUD instead of calling the function each time. There are probably more instances where this can be applied still, I've only done cl_hud.

Edited jcms.hud_GetLocatorColor to not use an extra if-statement. Made it return the color-value immediately instead of storing and returning it at the end.

Moved creation of matrices used in jcms.draw_Locators out of the function, so that that isn't done every frame unnecessarily.
Moved arrow shape definition out of the function (ditto)
Made the vector(s) used for applying transformations to the main matrix defined outside of the function, and replaced the new vector creations with :SetUnpacked()
Made isvector for loc.at only run if it isn't an entity, eliminated nothing-operation (setting it to the value we already know it has).

Removed circles from locators as these seem to be near-unnoticeable / obscured by locator text. May be reverted if Mereki informs me they had a good reason to exist. Under the circumstances I tested though this was a 7 pixel difference (on a large monitor) for running a pretty expensive function twice.