A bunch of small gameplay changes & fixes
Evac now tracks and displays clerks recruited as text (the render function for evac probably needs a refactor at some point, as it's inefficient and has duplicate code, which I have just made worse by nabbing Mereki's text drawing stuff). This should hopefully stop people from over-evacing clerks and stalling rounds as a result.

Evac now tracks if all sweepers are present and gets higher pitched if they are (and if the game is multiplayer). This isn't a good solution, but it's a fast one, and I don't want to spend time on a text for it yet.

Reduced liability threshold from 5 kills to 4

Made NPC players' damage follow NPC damage scaling (should make NPC players much more threatening on long runs, so that they match the danger posed by normal NPCs)

Saving runprogress data permanently (/until overwritten) if it's a high score. I want to use this to display the server's highest winstreak (and maybe the players that achieved it) in the future, to give people something to work towards.

Maybe fix for engine-error prevention code lua-erroring? This *shouldn't* fix it, but I'm working under the assumption something's wrong with my thinking and hoping it is just a weird edge value thing (The times it has shown up have implied it probably is, even though that doesn't make sense to me).

Made boss HP scaling taper off as the values start to get ridiculously high. This should be broadly the same overall, but values on the really high end will be significantly lower (e.g. 10x winstreak w/10 players was 15x HP before, now 7.5x, which is still a lot but more manageable).