Minimum sweepers to activate nuke, fix issue re-introduced in other commit
Made it so that nukes require 25% of sweepers (rounded up) to be present in-order to activate. This should make it less likely for idiots to immediately detonate them on public servers.

Fixed an issue re-introduced by a previous commit where objectives w/terminals on them could be spawned inside walls, making the mission impossible to complete.