Remove antlions if burrowed >60s
Antlion enemies still seem to be able to remain indefinitely burrowed in-spite of all of the checks and fall-backs to prevent this. This is a final fall-back that ensures they won't occupy soft-cap or block evac, at the expense of the enemy being lost (though it wouldn't have been able to spawn properly anyway).