Fixed map thumbnails (&another lua error?)
Moved map icons to the thumbnail directory, so that they can now actually be uploaded to the workshop properly.

Changed the player.GetCount check in the engine-error prevention hook to just check if the node at eIndex exists. This is honestly better overall and should hopefully finally fix the issue with that erroring (still don't know how it's getting into that part of the code without nodes existing. I was under the impression that eIndex for players always starts at / increases from 1, but maybe I'm wrong.)