Tracking valid maps
Valid maps are now stored in a table and saved on map reload. This list is checked-against first before the usual checks for navmesh/nodegraph.

This is a bandaid to general reliability issues with detecting navmesh and nodegraph files (Seemingly several unrelated issues. One is packed content, but there's also an issue where nodegraphs will just not exist sometimes. Even on the maps that come with the addon / even controlling for the cases in which you'd normally expect that to happen.)

I have hardcoded the recommended and base-game maps to be detected as valid to save people some trouble. gm_born in particular doesn't show up by default due to having packed content, but jcorpdistrict and mafiadistrict have occasionally had issues too for some reason.

New maps are added automatically when booted for the first time, if (via more reliable in-game methods) they're detected as valid.

Maps can be added manually via the jcms_addValidMap concmd, or by editing the .json