PLAYERCLIP DETECTION!
Data read from the map BSP file + a good bit of mathematics are now used to detect whether or not a nav-area intersects a player-clip. This algorithm probably isn't as refined / efficient as it could be, but it works. Which is the important part.

The code for setting up zones now uses validArea instead of doing its own separate duplicate validity checks. This won't affect performance (and may even improve it) as area validity is cached and has been for a while now.

The end-result of this is that objectives, players, and any other mission-critical objects should no longer spawn in areas that are blocked / inaccessible due to clips. Meaning many more maps should now be playable.

Also slipped a fix to rgg-grunts falling under displacements in as well.